# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ticket Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contacts
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: organizationId
          in: query
          schema:
            type: ["integer", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
      responses:
        "200":
          description: Contact list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactPage"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /contacts/{id}/tickets:
    get:
      operationId: listContactTickets
      tags: [contacts]
      summary: List tickets for a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"

  # ─── ORGANIZATIONS ────────────────────────────────────────────
  /organizations:
    get:
      operationId: listOrganizations
      tags: [organizations]
      summary: List organizations
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
      responses:
        "200":
          description: Organization list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationPage"
    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Create an organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationInput"
      responses:
        "201":
          description: Created organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /organizations/{id}:
    get:
      operationId: getOrganization
      tags: [organizations]
      summary: Get an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          description: Not found
    patch:
      operationId: updateOrganization
      tags: [organizations]
      summary: Update an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationUpdate"
      responses:
        "200":
          description: Updated organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    delete:
      operationId: deleteOrganization
      tags: [organizations]
      summary: Delete an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── AGENTS ───────────────────────────────────────────────────
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List agents
      responses:
        "200":
          description: Agent list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create an agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentInput"
      responses:
        "201":
          description: Created agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    patch:
      operationId: updateAgent
      tags: [agents]
      summary: Update an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentUpdate"
      responses:
        "200":
          description: Updated agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"

  # ─── TAGS ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    Ticket:
      type: object
      required: [id, subject, status, priority, type, channel, createdAt, updatedAt]
      properties:
        id:
          type: integer
        subject:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, pending, on_hold, solved, closed]
        priority:
          type: string
          enum: [low, normal, high, urgent]
        type:
          type: string
          enum: [question, incident, problem, task]
        channel:
          type: string
          enum: [email, chat, phone, web, api]
        assigneeId:
          type: ["integer", "null"]
        assignee:
          $ref: "#/components/schemas/Agent"
        requesterId:
          type: ["integer", "null"]
        requester:
          $ref: "#/components/schemas/Contact"
        organizationId:
          type: ["integer", "null"]
        organization:
          $ref: "#/components/schemas/Organization"
        tags:
          type: array
          items:
            type: string
        dueAt:
          type: ["string", "null"]
        firstResponseAt:
          type: ["string", "null"]
        resolvedAt:
          type: ["string", "null"]
        satisfaction:
          type: ["string", "null"]
          enum: [good, bad, null]
        commentCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Contact:
      type: object
      required: [id, name, email, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        organizationId:
          type: ["integer", "null"]
        organization:
          $ref: "#/components/schemas/Organization"
        role:
          type: string
          enum: [end_user, agent, admin]
        tags:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        ticketCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    ContactInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        organizationId:
          type: ["integer", "null"]
        role:
          type: string
          enum: [end_user, agent, admin]
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
    ContactUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        organizationId:
          type: ["integer", "null"]
        role:
          type: string
          enum: [end_user, agent, admin]
        tags:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
    ContactPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── ORGANIZATION ─────────────────────────────────────────
    Organization:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        domain:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        plan:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        contactCount:
          type: integer
        ticketCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    OrganizationInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        domain:
          type: string
        industry:
          type: string
        plan:
          type: string
        notes:
          type: string
    OrganizationUpdate:
      type: object
      properties:
        name:
          type: string
        domain:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        plan:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    OrganizationPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Organization"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── AGENT ────────────────────────────────────────────────
    Agent:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [agent, admin, supervisor]
        avatarUrl:
          type: ["string", "null"]
        isOnline:
          type: boolean
        openTicketCount:
          type: integer
        createdAt:
          type: string
    AgentInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [agent, admin, supervisor]
    AgentUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [agent, admin, supervisor]
        isOnline:
          type: boolean

    # ─── COMMENT ──────────────────────────────────────────────
...