# lib/api-zod/src/generated/types/agentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ticket Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentUpdateRole } from './agentUpdateRole';

export interface AgentUpdate {
  name?: string;
  email?: string;
  role?: AgentUpdateRole;
  isOnline?: boolean;
}