# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ticket Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { ContactInputRole } from './contactInputRole';

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  phone?: string;
  /** @nullable */
  organizationId?: number | null;
  role?: ContactInputRole;
  tags?: string[];
  notes?: string;
}